Extend test cert validity to 2049
to make tests pass in 2022

Background:
As part of my work on reproducible builds for openSUSE, I check
that software still gives identical build results in the future.
The usual offset is +15 years, because that is how long I expect
some software will be used in some places.
This showed up failing tests in our package build.

See https://reproducible-builds.org/ for why this matters.